refactor: :fire: remove and  fix paths
