--- src/cmsFaker/projects/variables.ts
@@ -1,2 +1,0 @@
-export const SEO_GITHUB = "github",
-  SEO_LINK = "link", SEO_FIGMA = "figma"
