+++ src/interface/Route.ts
@@ -1,0 +1,8 @@
+import { Language } from "./Lenguages";
+
+export interface IRoutes {
+  path: string;
+  title: string;
+  language?:Language;
+   __typename?:string;
+}
